clean some, osmo lp
