+++ indexer/cosmos.go
@@ -58,41 +58,7 @@ func NewCosmosIndexer(params CosmosIndexerParams) (*CosmosIndexer, error) {
 	return &CosmosIndexer{db: d, tm: tm, lcd: lcd, chain: chain, startHeight: params.StartHeight, endHeight: params.EndHeight}, nil
 }
 
-func (c *CosmosIndexer) IndexLP(poolName string) error {
-	startHeight := int64(c.chain.SnapshotStartBlock)
-	endHeight := int64(c.chain.SnapshotEndBlock)
-
-	latestBlock, err := c.tm.Block(context.Background(), nil)
-	if err != nil {
-		return errors.Wrapf(err, "error getting latest block")
-	}
-
-	if endHeight > latestBlock.Block.Height {
-		endHeight = latestBlock.Block.Height
-	}
-
-	latest, err := c.db.FindLatestIndexedThorchainLPBlock(c.chain.Name, poolName)
-	if err != nil {
-		return errors.Wrapf(err, "error finding latest indexed block")
-	}
-	if latest > startHeight {
-		log.Infof("found latest indexed block %d, starting at %d", latest, latest-1)
-		startHeight = latest - 1
-	}
-
-	blocksPerDay := int64(6 * 60 * 24)
-	// get up to an extra day of blocks, trim if post cutoff
-	for i := startHeight; i <= endHeight+blocksPerDay; i += blocksPerDay {
-		if err := c.indexLP(i, poolName); err != nil {
-			log.Errorf("error indexing delegations at height %d: %+v", i, err)
-		}
-		log.Infof("indexed %s lp block %d", poolName, i)
-	}
-
-	return nil
-}
-
-func (c *CosmosIndexer) IndexDelegators() error {
+func (c *CosmosIndexer) IndexCosmosDelegators() error {
 	startHeight := int64(c.chain.SnapshotStartBlock)
 	endHeight := int64(c.chain.SnapshotEndBlock)
 
@@ -106,7 +72,7 @@ func (c *CosmosIndexer) IndexDelegators() error {
 	}
 
 	for i := startHeight; i <= endHeight; i++ {
-		if err := c.indexDelegations(i); err != nil {
+		if err := c.indexCosmosDelegations(i); err != nil {
 			log.Errorf("error indexing delegations at height %d: %+v", i, err)
 		}
 	}
@@ -246,102 +212,7 @@ func attributesToMap(attributes []abcitypes.EventAttribute) map[string]string {
 	return m
 }
 
-type ThorLiquidityProvider struct {
-	Asset             string `json:"asset"`
-	AddressThor       string `json:"rune_address"`
-	AddressNative     string `json:"asset_address"`
-	LastAddHeight     int64  `json:"last_add_height"`
-	Units             int64  `json:"units,string"`
-	PendingRune       int64  `json:"pending_rune,string"`
-	PendingAsset      int64  `json:"pending_asset,string"`
-	RuneDepositValue  int64  `json:"rune_deposit_value,string"`
-	AssetDepositValue int64  `json:"asset_deposit_value,string"`
-}
-
-func (c *CosmosIndexer) findThorLpsByHeight(height int64, poolName string) ([]ThorLiquidityProvider, error) {
-	///thorchain/pool/ETH.FOX-0XC770EEFAD204B5180DF6A14EE197D99D808EE52D/liquidity_providers
-	path := fmt.Sprintf("/thorchain/pool/%s/liquidity_providers", poolName)
-	res, err := c.lcd.R().SetQueryParam("height", fmt.Sprintf("%d", height)).Get(path)
-	if err != nil {
-		return nil, errors.Wrapf(err, "error reading balances at height %d, pool %s", height, poolName)
-	}
-	if res.StatusCode() != 200 {
-		return nil, fmt.Errorf("error reading balances at height %d, pool %s, status code %d", height, poolName, res.StatusCode())
-	}
-
-	results := make([]ThorLiquidityProvider, 0, 1024)
-	if err = json.Unmarshal(res.Body(), &results); err != nil {
-		return nil, errors.Wrapf(err, "error unmarshalling balances at height %d, pool %s", height, poolName)
-	}
-	return results, nil
-}
-
-type ThorPool struct {
-	Asset               string `json:"asset"`
-	BalanceAsset        int64  `json:"balance_asset,string"`
-	BalanceRune         int64  `json:"balance_rune,string"`
-	PoolUnits           int64  `json:"pool_units,string"`
-	LpUnits             int64  `json:"LP_units,string"`
-	Status              string `json:"status"`
-	SynthSupply         int64  `json:"synth_supply,string"`
-	SynthUnits          int64  `json:"synth_units,string"`
-	PendingInboundRune  int64  `json:"pending_inbound_rune,string"`
-	PendingInboundAsset int64  `json:"pending_inbound_asset,string"`
-}
-
-func (c *CosmosIndexer) findPoolByHeight(height int64, poolName string) (ThorPool, error) {
-	pool := ThorPool{}
-	path := fmt.Sprintf("/thorchain/pool/%s", poolName)
-	res, err := c.lcd.R().SetQueryParam("height", fmt.Sprintf("%d", height)).Get(path)
-	if err != nil {
-		return pool, errors.Wrapf(err, "error reading pool at height %d, pool %s", height, poolName)
-	}
-	if res.StatusCode() != 200 {
-		return pool, fmt.Errorf("error reading pool at height %d, pool %s, status code %d", height, poolName, res.StatusCode())
-	}
-
-	if err = json.Unmarshal(res.Body(), &pool); err != nil {
-		return pool, errors.Wrapf(err, "error unmarshalling pool at height %d, pool %s", height, poolName)
-	}
-	return pool, nil
-}
-
-func (c *CosmosIndexer) indexLP(height int64, poolName string) error {
-	pool, err := c.findPoolByHeight(height, poolName)
-	if err != nil {
-		return errors.Wrapf(err, "error reading pool at height %d for %s", height, poolName)
-	}
-	_ = pool
-	totalUnits := pool.PoolUnits
-	lpBalances, err := c.findThorLpsByHeight(height, poolName)
-	if err != nil {
-		return errors.Wrapf(err, "error reading balances at height %d for %s", height, poolName)
-	}
-
-	batch := make([]types.ThorLPBalanceEvent, 0, len(lpBalances))
-	for _, bal := range lpBalances {
-		share := float64(bal.Units) / float64(totalUnits)
-		shareAsset := share * utils.FromBaseUnits(pool.BalanceAsset, 8)
-		shareRune := share * utils.FromBaseUnits(pool.BalanceRune, 8)
-		batch = append(batch,
-			types.ThorLPBalanceEvent{
-				Chain:         c.chain.Name,
-				BlockNumber:   int64(height),
-				Pool:          poolName,
-				AddressThor:   bal.AddressThor,
-				AddressNative: bal.AddressNative,
-				BalanceRune:   shareRune,
-				BalanceAsset:  shareAsset,
-			},
-		)
-	}
-	if err = c.db.InsertThorLPBalanceEvent(batch); err != nil {
-		return errors.Wrapf(err, "error inserting thor lp balances at height %d for %s", height, poolName)
-	}
-	return nil
-}
-
-func (c *CosmosIndexer) indexDelegations(height int64) error {
+func (c *CosmosIndexer) indexCosmosDelegations(height int64) error {
 	log := log.WithField("height", fmt.Sprintf("%d", height))
 	var (
 		ctx             = context.Background()
@@ -410,6 +281,25 @@ func parseAmount(in string, decimals uint8) (asset string, amount float64, err e
 
 	amount = utils.BigIntToFloat(iamt, uint8(decimals))
 	return
+}
+
+func (c *CosmosIndexer) indexOsmoLP(height int64) error {
+	log := log.WithField("height", fmt.Sprintf("%d", height))
+	var (
+		ctx = context.Background()
+		// txSearchResults []*coretypes.ResultTx
+		// txSearchErr     error
+	)
+	_ = log
+	// end block events have LP events for chains other than THOR
+	blockResults, err := c.tm.BlockResults(ctx, &height)
+	if err != nil {
+		return errors.Wrapf(err, "error reading search results height %d", height)
+	}
+	for _, evt := range blockResults.EndBlockEvents {
+		log.Infof("evt %s", evt.Type)
+	}
+	return nil
 }
 
 /*
